+++ nodes/input/input.length/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = (NODE) => {
+  const anyIn = NODE.getInputByName('any');
+
+  const lengthOut = NODE.getOutputByName('length');
+  lengthOut.on('trigger', async (conn, state) =>
+    (await anyIn.getValues(state)).length
+  );
+};